+++ lib/apartment.rb
@@ -4,5 +4,11 @@ require 'apartment/database'
 
 module Apartment
 
+  autoload :ApartmentError, 'apartment/errors'
+  autoload :Database, 'aparment/database'
+  
+  module Adapters
+    autoload AbstractAdapter, 'apartment/adapters/abstract_adapter'
+  end
 end