+++ lib/apartment/adapters/abstract_adapter.rb
@@ -1,0 +1,6 @@
+module Apartment
+  
+  class AbstractAdapter
+  end
+  
+end