+++ lib/apartment/adapters/mysql_adapter.rb
@@ -1,0 +1,6 @@
+module Apartment
+  
+  class MysqlAdapter < AbstractAdapter
+  end
+  
+end