+++ lib/apartment/adapters/postgresql_adapter.rb
@@ -1,0 +1,6 @@
+module Apartment
+  
+  class PostgresqlAdapter < AbstractAdapter
+  end
+  
+end